moves encryption scripts to separate file
